Create Linear Model
Important Statistical Tools: sklearn, imblearn, and SMOTE. 


(A Pharaoh, TensorFlow/Pytorch, and StatsModels/Modeler stack could be viable for this task as well depending on the use case and needed result). 



The first step is to import the necessary libraries to create a linear model to get a better understanding of the metrics we are working with!