Create Ensemble Methods
It never hurts to try other methods! We will now utilize a Stacking Ensemble Method to create a more intelligent nueral network that will decide what tests will be most effective for the data contained in our data set and run them! 


In order to do this we will need to make sure to import all the neccessary libraries to train, test, and split our machine learning algorithim for optimal model selection.